# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anony-mail API",
    "version": "0.3.0",
    "summary": "REST + Server-Sent Events API for an inbound-only disposable email backend.",
...
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "contact": {
      "name": "anony-mail source & issues",
      "url": "https://github.com/1mrnewton/anony-mail"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development (default API_BIND_ADDR)"
    },
    {
      "url": "https://{host}",
      "description": "Deployed instance",
      "variables": {
        "host": {
          "default": "api.example.com",
          "description": "Hostname where your anony-mail instance is reachable."
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/client/challenge": {
      "post": {
        "operationId": "clientChallenge",
        "summary": "Issue an App Attest challenge",
        "description": "Returns a single-use challenge for the attest/assert flows. Feed its SHA-256 (of the UTF-8 string) to `DCAppAttestService` as the `clientDataHash`, then send the same string back in the follow-up call. A challenge can be consumed exactly once and expires after `expires_in` seconds — request a fresh one per attempt.\n\nAvailable whenever App Attest is configured (`APP_ATTEST_TEAM_ID` + `APP_ATTEST_BUNDLE_ID` + `TOKEN_SIGNING_KEY`); 503 otherwise. Shares the strict address-creation rate-limit budget. No request body.",
        "tags": [
          "Client attestation"
        ],
        "responses": {
          "200": {
            "description": "A fresh challenge.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponse"
                },
                "example": {
                  "challenge": "vN3kx4nyBIRotmxTZ7qXW1ZQ8Kk5cLVX9y0mUuxbcpM",
                  "expires_in": 300
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/AttestationNotConfigured"
          }
        }
      }
    },
    "/api/client/attest": {
      "post": {
        "operationId": "clientAttest",
        "summary": "Attest a device key and mint an attested token",
        "description": "One-time device registration. Verifies the App Attest attestation object locally (certificate chain to Apple's App Attest root CA, app id, challenge hash), stores the device's public key, and returns an **attested** client token — same shape as `POST /api/entitlements/verify`, but it also satisfies servers with `CLIENT_ATTESTATION_REQUIRED=true`. Include the anonymous RevenueCat app-user id so the token carries the right tier.\n\nRun once per app install with the result of `DCAppAttestService.generateKey` + `attestKey`; afterwards refresh tokens with `POST /api/client/assert`. Rejections are coded (`400`): `invalid_challenge` for an unknown, expired, or reused challenge, `invalid_attestation` when verification fails.",
        "tags": [
          "Client attestation"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttestRequest"
              },
              "example": {
                "key_id": "Zju0hI0hL0iF2F6cx3QF6h0h4Y9V0kEEY0eXjxKq04c=",
                "attestation": "o2NmbXRvYXBwbGUtYXBwYXR0ZXN0Z2F0dFN0bXSiY3g1Y4JZAz...",
                "challenge": "vN3kx4nyBIRotmxTZ7qXW1ZQ8Kk5cLVX9y0mUuxbcpM",
                "rc_app_user_id": "$RCAnonymousID:87c6049c58069238dce29853f60e2c62"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device registered; attested token minted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientTokenResponse"
                },
                "example": {
                  "client_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJraW5kIjoiaW9zIiwia2lkIjoiWmp1MGhJMGhMMGlGMkY2Y3gzUUY2aDBoNFk5VjBrRUVZMGVYanhLcTA0Yz0iLCJ0aWVyIjoicHJvIn0.signature",
                  "tier": "pro",
                  "expires_at": "2026-08-22T19:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "The challenge or the attestation was rejected — `code` says which: `invalid_challenge` (request a new challenge and retry) or `invalid_attestation` (the attestation object failed verification).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "challenge unknown, expired, or already used; request a new one",
                  "code": "invalid_challenge"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/AttestationNotConfigured"
          }
        }
      }
    },
    "/api/client/assert": {
      "post": {
        "operationId": "clientAssert",
        "summary": "Refresh an attested token with an assertion",
        "description": "Refreshes the attested client token by proving possession of a previously registered device key: call `DCAppAttestService.generateAssertion` over the SHA-256 of a fresh challenge and send the result. The assertion's signature counter must strictly increase, which blocks replay — on a counter conflict just retry with a fresh challenge.\n\nCall it whenever the current token nears `expires_at` (the tier is re-checked on every refresh). Coded rejections: `invalid_challenge` and `invalid_assertion` (`400`), plus `attestation_required` (`401`) when the key id is unknown — for example pruned after long inactivity — in which case run the attest flow again.",
        "tags": [
          "Client attestation"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssertRequest"
              },
              "example": {
                "key_id": "Zju0hI0hL0iF2F6cx3QF6h0h4Y9V0kEEY0eXjxKq04c=",
                "assertion": "omlzaWduYXR1cmVYRzBFAiEA0YQx3Zt0iF2F6cx3QF6h0h4...",
                "challenge": "kQ8mZ1JXW1ZQ8Kk5cLVX9y0mUuxbcpMvN3kx4nyBIRo",
                "rc_app_user_id": "$RCAnonymousID:87c6049c58069238dce29853f60e2c62"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assertion verified; fresh attested token minted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientTokenResponse"
                },
                "example": {
                  "client_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJraW5kIjoiaW9zIiwia2lkIjoiWmp1MGhJMGhMMGlGMkY2Y3gzUUY2aDBoNFk5VjBrRUVZMGVYanhLcTA0Yz0iLCJ0aWVyIjoicHJvIn0.signature",
                  "tier": "pro",
                  "expires_at": "2026-08-22T19:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "The challenge or the assertion was rejected — `code` says which: `invalid_challenge` (request a new challenge and retry) or `invalid_assertion` (bad signature, wrong app id, or a non-increasing counter).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "assertion counter conflict, retry with a fresh challenge",
                  "code": "invalid_assertion"
                }
              }
            }
          },
          "401": {
            "description": "The key id is not registered on this server (never attested, or pruned after long inactivity). Run `POST /api/client/attest` again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "unknown key id, attest this device first",
                  "code": "attestation_required"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/AttestationNotConfigured"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ownerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-mailbox owner token returned once by `POST /api/addresses` (and by rotate). Format: `am_` + 43 base64url characters. Sent as `Authorization: Bearer <owner_token>`."
      },
      "claimToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-domain claim token returned once by `POST /api/custom-domains`. Format: `amd_` + 43 base64url characters. Sent as `Authorization: Bearer <claim_token>`."
      },
      "clientToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Token",
        "description": "Signed tier token (12h TTL) from `POST /api/entitlements/verify` (unattested) or `POST /api/client/attest` / `assert` (attested, App Attest). Optional: on enforcing servers a request without it is treated as free tier — unless the server reports `attestation.required`, in which case mutations demand an attested one. `Authorization` stays reserved for owner/claim tokens, so both can be sent together."
      }
    },
...
    "responses": {
...
      "TooManyRequests": {
        "description": "A per-IP rate limit or quota was exceeded. Retry later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "daily address creation limit reached"
            }
          }
        }
      },
      "AttestationNotConfigured": {
        "description": "App Attest is not configured on this server (`APP_ATTEST_TEAM_ID` / `APP_ATTEST_BUNDLE_ID` / `TOKEN_SIGNING_KEY` unset). Check `GET /api/capabilities` → `attestation.supported` before calling.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "app attestation is not configured on this server"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An unexpected server error occurred. Details are logged server-side.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "internal server error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "description": "Uniform error envelope returned on any non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message.",
            "examples": [
              "mailbox not found"
            ]
          },
          "code": {
            "type": "string",
            "description": "Optional machine-readable rejection code, present on policy errors: `pro_required`, `lifetime_cap`, `client_token_expired`, `client_token_invalid`. More codes may be added; ignore unknown ones.",
            "examples": [
              "pro_required"
            ]
          }
        }
      },
...
      "ClientTokenResponse": {
        "type": "object",
        "description": "A freshly minted client token — from `POST /api/entitlements/verify` (unattested) or the attest/assert endpoints (attested). Send it as `X-Client-Token` on writes; refresh via the endpoint that minted it when it nears `expires_at`.",
        "required": [
          "client_token",
          "tier",
          "expires_at"
        ],
        "properties": {
          "client_token": {
            "type": "string",
            "description": "Signed tier token (JWT, HS256). Attested tokens additionally carry the device key id."
          },
          "tier": {
            "type": "string",
            "enum": [
              "free",
              "pro"
            ],
            "description": "The tier embedded in the token."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the token stops validating (~12h)."
          }
        }
      },
      "ChallengeResponse": {
        "type": "object",
        "description": "A single-use App Attest challenge.",
        "required": [
          "challenge",
          "expires_in"
        ],
        "properties": {
          "challenge": {
            "type": "string",
            "description": "Opaque base64url string. Hash it (SHA-256 of the UTF-8 bytes) into the `clientDataHash` for `DCAppAttestService`, and echo it verbatim in the attest/assert request."
          },
          "expires_in": {
            "type": "integer",
            "description": "Seconds until the challenge expires.",
            "examples": [
              300
            ]
          }
        }
      },
      "AttestRequest": {
        "type": "object",
        "description": "One-time App Attest device registration.",
        "required": [
          "key_id",
          "attestation",
          "challenge"
        ],
        "properties": {
          "key_id": {
            "type": "string",
            "description": "Base64 key id returned by `DCAppAttestService.generateKey`."
          },
          "attestation": {
            "type": "string",
            "description": "Base64-encoded CBOR attestation object from `attestKey(_:clientDataHash:)`."
          },
          "challenge": {
            "type": "string",
            "description": "The challenge (verbatim) whose SHA-256 was used as `clientDataHash`."
          },
          "rc_app_user_id": {
            "type": "string",
            "description": "Optional anonymous RevenueCat app-user id for the tier check; omit for a free-tier token."
          }
        }
      },
      "AssertRequest": {
        "type": "object",
        "description": "Assertion-based refresh for an already-attested device.",
        "required": [
          "key_id",
          "assertion",
          "challenge"
        ],
        "properties": {
          "key_id": {
            "type": "string",
            "description": "The key id registered via `POST /api/client/attest`."
          },
          "assertion": {
            "type": "string",
            "description": "Base64-encoded CBOR assertion from `generateAssertion(_:clientDataHash:)`."
          },
          "challenge": {
            "type": "string",
            "description": "The challenge (verbatim) whose SHA-256 was used as `clientDataHash`."
          },
          "rc_app_user_id": {
            "type": "string",
            "description": "Optional anonymous RevenueCat app-user id for the tier check; omit for a free-tier token."
          }
        }
      }
    }
  }
}